fix(go-api): sync with CRD changes (Option or serde_default in Rust)
NOTE: `omitempty` or `pointer` are generated. Don't modify these
keywords manullay unless you are clear the effect.

The field marked with `Option` in rust will be `nullable` and not
required in API schema, and in generated golang code, the field is a
`Pointer`. While the field marked with `serde_default` in rust will
be `not nullable` and not required in API schema and in generated
golang code, the field is not a pointer but marked with `omitempty`.